Add .gitignore, ResponseException and abort()
ResponseException is a flow control exception (it inherits from
BaseException and not Exception) that is caught by the response
generator. This helps clean up control flow by avoiding having to pass
responses back.

This mechanism is used by abort(), which takes a 4xx or 5xx status code
and returns it to the client immediately.